% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{BCVdata}
\alias{BCVdata}
\title{Event study estimates from Bosch and Campos-Vazquez (2014)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Bosch and Campos-Vazquez (2014)
}
\references{
Bosch, M. and Campos-Vazquez, R.M. (2014). The Trade-Offs of Welfare Policies in Labor Markets with Informal Jobs: The Case of the "Seguro Popular" Program in Mexico. American Economic Journal: Economic Policy 6(4), 71–99
}
