% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{CCYdata}
\alias{CCYdata}
\title{Event study estimates from Chen et al. (2025)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Chen et al. (2025)
}
\references{
Chen, H., Chen, Y. and Yang, Q. (2025). Women in the courtroom: Technology and Justice. The Review of Economic Studies, 1–28.
}
