% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{CDGSWdata}
\alias{CDGSWdata}
\title{Event study estimates from Coskun et al. (2026)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Coskun et al. (2026)
}
\references{
Coskun, S., Dauth, W., Gartner, H., Stops, M. and Weber, E. (2026). Working from home increases work–home distances. Journal of Urban Economics 152, 103832
}
