% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2tau_fun2.R
\name{cov2tau_fun2}
\alias{cov2tau_fun2}
\title{This function computes the estimate of roughness parameter function tau(t) using the covariance function of the functional data}
\usage{
cov2tau_fun2(cov_mat, grid.min, grid.max, warn = FALSE)
}
\arguments{
\item{cov_mat}{matrix (pxp) of evaluated covariance function (p=number of discretization point).}

\item{grid.min}{min value of grid.}

\item{grid.max}{max value of grid.}

\item{warn}{option for printing warnings.}
}
\value{
estimate of the roughness parameter function tau(t).
}
\description{
This function computes the estimate of roughness parameter function tau(t) using the covariance function of the functional data.
This is adapted from the \link[ffscb]{cov2tau_fun} in package \pkg{ffscb}, but the range of grid is not limited to \link{0,1}.
}
\references{
Liebl, D. and M. Reimherr (2023). Fast and fair simultaneous confidence bands for functional parameters. Journal of
the Royal Statistical Society Series B: Statistical Methodology 85(3), 842–868
}
\seealso{
\link[ffscb]{cov2tau_fun}
}
