% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_intrpl.R
\name{cov_intrpl}
\alias{cov_intrpl}
\title{Natural Spline Interpolation on Covariance}
\usage{
cov_intrpl(cov, grid, n_intrpl)
}
\arguments{
\item{cov}{a numeric square matrix of raw covariance. It must not contain missing values.}

\item{grid}{a numeric vector of grid, at which the raw covariance is calculated.}

\item{n_intrpl}{the number of points at which the covariance is interpolated. The interpolation is performed on a sequence
of \code{n_intrpl} equidistant points from the minimum to the maximum of \code{grid}.}
}
\value{
The \code{cov_intrpl} function returns a numeric matrix of covariance interpolation.
}
\description{
The \code{cov_intrpl} function is used to perform natural spline interpolation on the covariance matrix.
See Fang and Liebl(2025) for detailed information.
}
\examples{
cov_mat <- cov(matrix(rnorm(250),nrow=50))
grid <- seq(0,1,len=5)
cov_intrpl_mat <- cov_intrpl(cov=cov_mat, grid=grid, n_intrpl=50)
}
\references{
Fang, C. and Liebl, D. (2025). Honest Causal Inference with Difference-in-Differences: A Functional Data Perspective.
}
