% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_spline.R
\name{cov_spline}
\alias{cov_spline}
\title{Natural spline interpolation on covariance}
\usage{
cov_spline(cov, grid, n_intrpl)
}
\arguments{
\item{cov}{a numeric square matrix of raw covariance. It must not contain missing values.}

\item{grid}{a numeric vector of grid, at which the raw covariance is calculated.}

\item{n_intrpl}{the number of points at which the covariance is interpolated. The interpolation is performed on a sequence
of \code{n_intrpl} equidistant points from the minimum to the maximum of \code{grid}.}
}
\value{
The \code{cov_spline} function returns a numeric matrix of covariance interpolation.
}
\description{
The \code{cov_spline} function is used to perform natural spline interpolation on the covariance matrix.
}
\examples{
cov_mat <- cov(matrix(rnorm(250),nrow=50))
grid <- seq(0,1,len=5)
cov_spline_mat <- cov_spline(cov=cov_mat, grid=grid, n_intrpl=50)
}
\references{
Fang, C. and Liebl, D. (2026). Making Event Study Plots Honest: A Functional Data Approach to Causal Inference. \href{https://arxiv.org/abs/2512.06804}{arXiv:2512.06804}.
}
