% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyPlot_Classical.R
\name{EventStudyPlot_Classical}
\alias{EventStudyPlot_Classical}
\title{Classical event study plot}
\usage{
EventStudyPlot_Classical(object, pos.legend = "top", scale.legend = 1, ...)
}
\arguments{
\item{object}{an object of class \code{"fdid_scb"}. The object to be plotted.}

\item{pos.legend}{a character value of "top" or "bottom" that indicates the position of legend. If NULL, the legend is not printed.}

\item{scale.legend}{a positive number that defines the size of legend. If \code{pos.legend} is NULL, the value of \code{scale.legend} is ignored.}

\item{...}{Additional arguments to be passed to \code{EventStudyPlot_Classical}.}
}
\value{
The \code{EventStudyPlot_Classical} function returns a classical event study plot with point-wise confidence intervals for event study coefficients.
}
\description{
This \code{EventStudyPlot_Classical} function is to draw the classical event study plot with point-wise confidence intervals for event study coefficients.
}
\examples{
data(LWdata)
fdid_scb_est <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)

par(cex.axis = 1.4, cex.lab = 1.4, cex.main = 1.4)
EventStudyPlot_Classical(fdid_scb_est, scale.legend=1.4)
title(ylab="Effects of Duty-to-Bargain Laws")
}
\references{
Fang, C. and Liebl, D. (2026). Making Event Study Plots Honest: A Functional Data Approach to Causal Inference. \href{https://arxiv.org/abs/2512.06804}{arXiv:2512.06804}.
}
\seealso{
\link{fdid_scb}
}
