% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdid_scb.R
\name{fdid_scb}
\alias{fdid_scb}
\title{Compute simultaneous confidence band for event study coefficients in functional DiD framework}
\usage{
fdid_scb(
  object = NULL,
  beta = NULL,
  cov = NULL,
  t0 = NULL,
  df = NULL,
  conf.level = 0.95
)
}
\arguments{
\item{object}{an object of S3 class \code{"fdid"}, which is an output of the function \link{fdid}. If it is provided, arguments 'beta',
'cov' and 't0' are ignored.}

\item{beta}{a numeric matrix with two columns. The first column is the estimates of event study coefficients and the second column is the
corresponding event time. The estimate of coefficient at reference time should be normalized to 0.}

\item{cov}{a numeric matrix of covariance estimates.}

\item{t0}{a numeric scalar that indicates the reference time.}

\item{df}{degree of freedom for the t-distributed based bands. If NULL, Gaussian distributed bands are computed.}

\item{conf.level}{confidence level for the bands to achieve.}
}
\value{
The \code{fdid_scb} function returns a list containing three objects:
\describe{
\item{scb}{a list that includes the functional estimate of event study coefficients and the simultaneous confidence band.}
\item{ci}{a matrix that includes the point-wise estimates of event study coefficients and their confidence intervals.}
\item{data}{a list that stores the original inputs of beta and cov.}
}
The output is an object of S3 class \code{"fdid_scb"}.
}
\description{
The \code{fdid_scb} function is used to compute simultaneous confidence band for event study coefficients in a functional
framework, together with the point-wise confidence intervals at observable event time. We adopt the algorithm of the so-called fast and
fair simultaneous confidence bands. See Liebl and Reimherr(2023) for details.
}
\examples{
data(LWdata)
fdid_scb_est1 <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)

data(simulated_stagger_example)
fdid_est2 <- fdid(data=simulated_stagger_data, treatment=simulated_stagger_treatment)
fdid_scb_est2 <- fdid_scb(object=fdid_est2)
}
\references{
Fang, C. and Liebl, D. (2025). Honest Causal Inference with Difference-in-Differences: A Functional Data Perspective.

Liebl, D. and M. Reimherr (2023). Fast and fair simultaneous confidence bands for functional parameters. Journal of
the Royal Statistical Society Series B: Statistical Methodology 85(3), 842–868
}
\seealso{
\link{fdid}, \link[ffscb]{confidence_band}, \link[ffscb]{confidence_band_fragm}, \link{cov_intrpl}
}
