% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Gdata}
\alias{Gdata}
\title{Event study estimates from Gallagher (2014)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Gallagher (2014)
}
\references{
Gallagher, J. (2014). Learning about an Infrequent Event: Evidence from Flood Insurance Take-Up in the United States. American Economic Journal: Applied Economics 6(3), 206–33.
}
