% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Ldata}
\alias{Ldata}
\title{Event study estimates from Lafortune et al. (2018)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Lafortune et al. (2018)
}
\references{
Lafortune, J., Rothstein, J. and Schanzenbach, D.W. (2018). School Finance Reform and the Distribution of Student Achievement. American Economic Journal: Applied Economics 10(2), 1–26.
}
