% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{LWdata}
\alias{LWdata}
\title{Event study estimates from Lovenheim and Willen (2019)}
\format{
A list, containing 4 objects:
\describe{
\item{beta}{Estimates of event study coefficients.}
\item{cov}{Estimates of covariance matrix.}
\item{t0}{Reference period of DiD design.}
\item{paper}{Paper indexing.}
}
}
\description{
Event study estimates from Lovenheim and Willen (2019)
}
\references{
Lovenheim, M.F. and Willen, A. (2019). The Long-Run Effects of Teacher Collective Bargaining. American Economic Journal: Economic Policy 11(3), 292– 324
}
