% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mortality}
\alias{mortality}
\title{FDID example dataset}
\format{
A data frame with X rows and Y columns.
}
\usage{
mortality
}
\description{
A long-format panel dataset for demonstrating the fdid package.
}
\keyword{datasets}
