% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdid_scb.R
\name{plot.fdid_scb}
\alias{plot.fdid_scb}
\title{Custom plot method for objects of class \code{"fdid_scb"}}
\usage{
\method{plot}{fdid_scb}(
  object,
  ta.ts = NULL,
  ta.s = NULL,
  ftr.m = NULL,
  ref.band.pre = TRUE,
  note.pre = TRUE,
  note.post = TRUE,
  ci.pre = FALSE,
  ci.post = FALSE,
  pos.legend = "top",
  scale.legend = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"fdid_scb"}. The object to be plotted.}

\item{ta.ts}{a numeric value that indicates the last time point in the pre-anticipation periods. It should be smaller than the reference time period.}

\item{ta.s}{a numeric vector of control parameters with length two for deriving the honest reference band under the violation of no-anticipation assumption. If \code{ta.ts} is NULL, the value of \code{ta.s} is ignored. If \code{ta.ts} is not NULL, the value of \code{ta.s} cannot be NULL.}

\item{ftr.m}{a numeric vector of control parameters with length two for deriving the honest reference band under violation of the parallel trends assumption. The default is NULL.}

\item{ref.band.pre}{a logical value. If TRUE, the reference band for pre-anticipation period is also plotted.}

\item{note.pre}{a logical value. If TRUE, the note for pre-anticipation period is given on top of plot. If no honest reference band is defined, there is no need to perform validation in the pre-anticipation period, and \code{note.pre} is FALSE.}

\item{note.post}{a logical value. If TRUE, the note for post-treatment period is given on top of plot.}

\item{ci.pre}{a logical value. If TRUE, the point-wise confidence intervals for pre-treatment period are plotted.}

\item{ci.post}{a logical value. If TRUE, the point-wise confidence intervals for post-treatment period are plotted.}

\item{pos.legend}{a character value of "top" or "bottom" that indicates the position of legend. If NULL, the legend is not printed.}

\item{scale.legend}{a positive number that defines the size of legend. If \code{pos.legend} is NULL, the value of \code{scale.legend} is ignored.}

\item{verbose}{a logical value. If TRUE, the uniformly significant time span(s) under your specification will be directly printed out on the console.}

\item{...}{Additional arguments to be passed to \link[base]{plot}.}
}
\value{
The function returns a plot with simultaneous confidence bands for event study coefficients in a functional framework, together with the honest reference band for honest inference, if properly defined.
}
\description{
This function is an S3 method for \link[base]{plot}, specifically designed for objects of class \code{"fdid_scb"}. It plots the simultaneous confidence bands together with honest reference band for performing honest causal inference.
}
\examples{
data(LWdata)
fdid_scb_est <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)
cat("The reference time is ", LWdata$t0, ". If not NULL, the input 'ta.ts' in function 'plot' should be smaller than this value.", sep="")

## simultaneous inference
par(cex.axis = 1.4, cex.lab = 1.4, cex.main = 1.4)
plot(fdid_scb_est, scale.legend=1.4)
title(ylab="Effects of Duty-to-Bargain Laws")

## honest inference under treatment anticipation
plot(fdid_scb_est, ta.ts=-3, ta.s=c(0.2,0.2), scale.legend=1.4)

## honest inference under violation of parallel trends assumption
plot(fdid_scb_est, ftr.m=c(0.2,0.2), scale.legend=1.4)
}
\references{
Fang, C. and Liebl, D. (2026). Making Event Study Plots Honest: A Functional Data Approach to Causal Inference. \href{https://arxiv.org/abs/2512.06804}{arXiv:2512.06804}.
}
\seealso{
\link{fdid_scb}
}
