% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fdid}
\alias{print.fdid}
\title{Print Method for FDID Objects}
\usage{
\method{print}{fdid}(x, ...)
}
\arguments{
\item{x}{An object of class `fdid`.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a brief overview of the `fdid` object
}
\description{
Print Method for FDID Objects
}
\author{
Rivka Lipkovitz.
}
