% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{stagger_example}
\alias{stagger_example}
\title{Simulated example of staggered DiD design}
\format{
It contains two data frames. The first data frame 'stagger_data' has 4200 rows and 5 columns:
\describe{
\item{y}{Outcome variable.}
\item{t}{Time index. There are 21 time periods.}
\item{i}{Unit index. There are 200 units.}
}
The second data frame 'stagger_treatment' has 200 rows and 2 columns:
\describe{
\item{i}{Unit index. It also corresponds to the unit index in data frame 'stagger_data'.}
\item{t0}{Reference Period. NA indicates control units.}
\item{x1, x2}{Pre-determined covariates.}
}
}
\description{
Simulated example of staggered DiD design
}
