% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fdid}
\alias{summary.fdid}
\title{Summary Method for FDID Objects}
\usage{
\method{summary}{fdid}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fdid}.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary of the \code{fdid} object.
}
\description{
Summary Method for FDID Objects
}
\examples{
\donttest{
data(fdid)
mortality$uniqueid <- paste(mortality$provid, mortality$countyid, sep = "-")
mortality$G <- ifelse(mortality$pczupu >= median(mortality$pczupu, na.rm = TRUE), 1, 0)
s <- fdid_prepare(
  data = mortality, Y_label = "mortality",
  X_labels = c("avggrain", "lnpop"),
  G_label = "G", unit_label = "uniqueid", time_label = "year"
)
result <- fdid(s, tr_period = 1958:1961, ref_period = 1957)
summary(result)
}
}
\author{
Rivka Lipkovitz, Enhan Liu
}
