% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_transf.R
\name{tw_transf}
\alias{tw_transf}
\title{Univariate two-way transformation}
\usage{
tw_transf(x, t, i)
}
\arguments{
\item{x}{a numeric vector the two-way transformation is supposed to be performed on.}

\item{t}{a vector of integers that indicates the time.}

\item{i}{a vector that indicates the unit.}
}
\value{
The \code{tw_transf} function returns two-way transformed data which are reordered by unit and time.
}
\description{
The \code{tw_transf} function is used to apply two-way transformation on a single variable.
}
\examples{
set.seed(100)
x <- rnorm(20)
t <- rep(1:5, each = 4)
i <- rep(letters[1:4], times=5)
df_transf <- tw_transf(x,t,i)
}
