% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdid.R
\name{fdid}
\alias{fdid}
\title{Estimate the event study coefficients and the covariance matrix}
\usage{
fdid(data, treatment)
}
\arguments{
\item{data}{a data frame in which the first variable should be the outcome variable, and the last two variables are time and unit indices. Covariates are placed
after the outcome but before time and unit variables. Outcome and covariates should be numeric.}

\item{treatment}{a data frame that indicates the reference time for each unit, i.e. the time right after which a treatment effect is present. NA implies that the
unit is never treated/control.}
}
\value{
The \code{fdid} function returns a list which includes the estimates of event study coefficients and their covariance. It also contains the estimates of
time-invariant coefficients for covariates and their standard errors at each period. In the output, the event time 0 is taken as the reference time. The output is
an object of S3 class \code{"fdid"}.
}
\description{
The \code{fdid} function is used to estimate the event study coefficients and the covariance matrix in non-staggered or staggered adoption design.
In staggered design, our estimation does not suffer from the so-called negative weighting problem, since we use carefully-chosen non-negative weights to consolidate
estimates from each subgroup. See Fang and Liebl(2025) for detailed estimation method.
}
\examples{
data(simulated_stagger_example)
fdid_est <- fdid(data=simulated_stagger_data, treatment=simulated_stagger_treatment)
plot(fdid_est$beta$coef[,2], fdid_est$beta$coef[,1], type="l",
     xlab="Event Time", ylab=expression(hat(beta)), family="Times")
}
\references{
Fang, C. and Liebl, D. (2025). Honest Causal Inference with Difference-in-Differences: A Functional Data Perspective.
}
\seealso{
\link{tw_transf}, \link{fdid_scb}
}
