% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fdid}
\alias{plot.fdid}
\title{Plot method for 'fdid' objects}
\usage{
\method{plot}{fdid}(
  x,
  type = c("raw", "dynamic", "overlap", "est"),
  alpha_shade = 0.2,
  col_treated = "red",
  col_control = "blue",
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class 'fdid', or a list of multiple 'fdid' objects
(if type='est').}

\item{type}{The type of plot to produce: "raw", "dynamic", "overlap", or "est".}

\item{alpha_shade}{Transparency level for shading the treatment period.}

\item{col_treated, col_control}{Colors for treated vs. control lines (raw means).}

\item{xlab, ylab, main}{Axis labels and main title.}

\item{ylim}{User-specified y-axis limits (optional).}

\item{...}{Further arguments passed to internal plotting functions.}
}
\value{
A base R plot is produced. Nothing is returned.
}
\description{
Plot method for 'fdid' objects
}
