% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdid_scb.R
\name{plot.fdid_scb}
\alias{plot.fdid_scb}
\title{Custom Plot Method for objects of class \code{"fdid_scb"}}
\usage{
\method{plot}{fdid_scb}(
  object,
  ci = TRUE,
  ta.t0 = 0,
  frm.mbar = NULL,
  ftr.m = NULL,
  legend = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"fdid_scb"}. The object to be plotted.}

\item{ci}{a logical value. If TRUE, the point-wise confidence intervals are also plotted.}

\item{ta.t0}{a numeric value that indicates the time, right after which there is an anticipation of treatment. It should be
greater than the minimal event time and no greater than 0. If it equals to 0, there is no anticipation.}

\item{frm.mbar}{a numeric value of tuning parameter for functional relative magnitudes.}

\item{ftr.m}{a numeric value of tuning parameter for functional trend restrictions.}

\item{legend}{a logical value. If TRUE, a legend for honest inference is printed on the top of plot.}

\item{...}{Additional arguments to be passed to \link[base]{plot}.}
}
\value{
The function returns a plot with simultaneous confidence band for event study coefficients in a functional framework,
together with bounds for honest inference, if properly defined.
}
\description{
This function is an S3 method for \link[base]{plot}, specifically designed for objects of class \code{"fdid_scb"}.
It plots simultaneous confidence band with bounds for honest inference.
}
\note{
The inference result around event time 0 (i.e. between two event time closest to 0) should be treated with caution.
}
\examples{
data(LWdata)
fdid_scb_est <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)
plot(fdid_scb_est)
plot(fdid_scb_est, frm.mbar=0.3)
}
\references{
Fang, C. and Liebl, D. (2025). Honest Causal Inference with Difference-in-Differences: A Functional Data Perspective.
}
\seealso{
\link{fdid_scb}
}
