% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdid_scb.R
\name{plot.fdid_scb}
\alias{plot.fdid_scb}
\title{Custom Plot Method for objects of class \code{"fdid_scb"}}
\usage{
\method{plot}{fdid_scb}(
  object,
  ta.t0 = NULL,
  ta.s = NULL,
  frm.mbar = NULL,
  ftr.m = NULL,
  frmtr.mbar = NULL,
  post.trt = FALSE,
  ci = TRUE,
  pos.legend = "top",
  ...
)
}
\arguments{
\item{object}{an object of class \code{"fdid_scb"}. The object to be plotted.}

\item{ta.t0}{a numeric value that indicates the time, right after which there is an anticipation of treatment. It should be
greater than the minimal event time and no greater than t0 in \code{object}. If it is NULL, there is no anticipation.}

\item{ta.s}{a numeric value of tuning parameter for honest inference under treatment anticipation. If NULL, the critical value \eqn{t_{1-\alpha/2, df}} or \eqn{z_{1-\alpha/2}} is applied with the same \eqn{\alpha} used for constructing simultaneous confidence band. If no treatment anticipation is defined in \code{ta.t0}, the value of \code{ta.s} is ignored.}

\item{frm.mbar}{a numeric value of tuning parameter for honest inference under functional relative magnitudes.}

\item{ftr.m}{a numeric value of tuning parameter for honest inference under functional trend restrictions.}

\item{frmtr.mbar}{a numeric value of tuning parameter for combining functional relative magnitudes and functional trend restrictions.}

\item{post.trt}{a logical value. If TRUE, only bounds for post-treatment periods are plotted and used for honest inference.}

\item{ci}{a logical value. If TRUE, the point-wise confidence intervals are also plotted.}

\item{pos.legend}{a character value of "top" or "bottom" that indicates the position of legend. If NULL, the legend is not printed.}

\item{...}{Additional arguments to be passed to \link[base]{plot}.}
}
\value{
The function returns a plot with simultaneous confidence band for event study coefficients in a functional framework,
together with bounds for honest inference, if properly defined.
}
\description{
This function is an S3 method for \link[base]{plot}, specifically designed for objects of class \code{"fdid_scb"}.
It plots simultaneous confidence band with bounds for honest inference.
}
\note{
The inference result around the reference time point (i.e. between two event time closest to the reference time) should be treated with caution.
}
\examples{
data(LWdata)
fdid_scb_est <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)
cat("The reference time is ", LWdata$t0, ". If not NULL, the input 'ta.t0' in function 'plot' should be smaller than this value.", sep="")

## simultaneous inference
plot(fdid_scb_est)

## honest inference under treatment anticipation
plot(fdid_scb_est, ta.t0=-4)

## honest inference under violation of parallel trends
plot(fdid_scb_est, frm.mbar=0.3)
}
\references{
Fang, C. and Liebl, D. (2025). Honest Causal Inference with Difference-in-Differences: A Functional Data Perspective.
}
\seealso{
\link{fdid_scb}
}
