% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fdid}
\alias{summary.fdid}
\title{Summary Method for FDID Objects}
\usage{
\method{summary}{fdid}(object, ...)
}
\arguments{
\item{object}{An object of class `fdid`.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary of the `fdid` object.
}
\description{
Summary Method for FDID Objects
}
\author{
Rivka Lipkovitz
}
